feat: upgrade to redwood and add deprecation notice (#23)
upgrade to redwood, add deprecation notice, migrate to github actions, remove python 2.7 support and remove test workflow

The tests depend on xblock-utils package which depends on some workbench
module that does not exists. Since the repository is planned to be
deprecated until we get real users, we don't want to spend time fixing
it.